fix: move question highligh to general dto
